added select to record card, matched update on cart and recordcard
